+++ components/builder-web/app/origin/origin-page/origin-keys-tab/key-add-form/key-add-form.component.html
@@ -1,0 +1,41 @@
+<form class="hab-key-add-form" [formGroup]="form" #formValues="ngForm"
+(ngSubmit)="submit(formValues.value.key)">
+  <a class="hab-key-add-form--close" href="#" (click)="onCloseClick()">
+      Close
+  </a>
+  <label class="hab-key-add-form--label" for="key">Key</label>
+  <small>
+    Paste your key here. Check the documentation for a guide on
+    <a href="{{docsUrl}}/share-packages-overview/">
+        generating keys</a>.
+  </small>
+  <textarea
+    autofocus
+    name="key"
+    [formControl]="form.controls['key']"
+    placeholder="Begins with '{{keyFileHeaderPrefix}}'"
+    rows=6>
+  </textarea>
+  <div class="hab-key-add-form--submit">
+    <button class="cta hab-key-add-form--save" [disabled]="!form.valid">
+        Upload Key
+    </button>
+    <div *ngIf="control.dirty && control.errors" class="hab-key-add-form--errors">
+      <span *ngIf="control.errors.required">
+        A value is required.
+      </span>
+      <span *ngIf="control.errors.invalidFormat">
+        This is not a valid key format.
+      </span>
+      <span *ngIf="control.errors.invalidType">
+        Key must begin with '{{keyFileHeaderPrefix}}'.
+      </span>
+      <span *ngIf="control.errors.invalidOrigin">
+        Key origin must match '{{originName}}'.
+      </span>
+    </div>
+    <div *ngIf="errorMessage" class="hab-key-add-form--errors">
+      Failed to save key: {{errorMessage}}.
+    </div>
+  </div>
+</form>
