+++ components/builder-web/app/origin/origin-page/origin-keys-tab/key-add-form/key-add-form.component.scss
@@ -1,0 +1,32 @@
+.hab-key-add-form {
+  @include shadow-float-more;
+  border-radius: $global-radius;
+  margin-bottom: rem(30);
+  padding: rem(20);
+
+  &--close {
+    float: right;
+  }
+
+  &--label {
+    color: $hab-gray-dark;
+    font-size: rem(16);
+    margin-top: $base-spacing;
+  }
+
+  &--submit {
+    @include outer-container;
+  }
+
+  &--save {
+    @include span-columns(4);
+    margin-top: 0;
+  }
+
+  &--errors {
+    @include span-columns(8);
+    @include omega;
+    color: $hab-red;
+    font-size: $error-font;
+  }
+}
