+++ components/builder-web/app/origin/origin-page/origin-keys-tab/key-list/key-list.component.html
@@ -1,0 +1,16 @@
+<p *ngIf="keys.size === 0">
+  No {{type}} keys found.
+</p>
+<p *ngIf="keys.size > 0 && publicKey">
+  Select any key to download the file.
+</p>
+<ul class="hab-key-list--list" *ngIf="keys.size > 0">
+  <li *ngFor="let key of keys" class="hab-item-list--all-link hab-item-list">
+    <h3>
+      <a href="{{apiUrl}}/depot{{key.location}}" *ngIf="publicKey">
+        {{key.origin}}-{{key.revision}}
+      </a>
+      <span *ngIf="!publicKey">{{key}}</span>
+    </h3>
+  </li>
+</ul>
