+++ components/builder-web/app/origin/origin-page/origin-keys-tab/key-list/key-list.component.scss
@@ -1,0 +1,5 @@
+.hab-key-list {
+  &--list {
+    margin-top: $base-spacing;
+  }
+}
