+++ components/builder-web/app/origin/origin-page/origin-packages-tab/origin-packages-tab.component.html
@@ -1,0 +1,46 @@
+<div class="page-body has-sidebar">
+  <div class="hab-origin--left hab-origin--pkg-list">
+    <div *ngIf="noPackages">
+      <p>
+        No packages found.
+        <span *ngIf="packagesUi.errorMessage">
+            Error: {{packagesUi.errorMessage}}
+        </span>
+      </p>
+    </div>
+
+    <div *ngIf="!noPackages">
+      <div class="pkg-container">
+        <div class="pkg-col-1">Package Name</div>
+        <div class="pkg-col-2">&nbsp;</div>
+        <div class="pkg-col-3">Versions</div>
+      </div>
+      <div *ngFor="let pkg of packages" class="pkg-container">
+        <div class="pkg-col-1">
+          <h3>{{pkg.name}}</h3>
+        </div>
+        <div class="pkg-col-2">
+          &nbsp;
+        </div>
+        <div class="pkg-col-3">
+          <a [routerLink]="['/pkgs', pkg.origin, pkg.name]">
+            <hab-icon symbol="layers"></hab-icon>
+          </a>
+        </div>
+      </div>
+      <div *ngIf="packages.size < totalCount">
+        Showing {{packages.size}} of {{totalCount}} packages.
+        <a href="#" (click)="fetchMorePackages()">
+          Load
+          {{(totalCount - packages.size) > perPage ? perPage : totalCount - packages.size }}
+          more</a>.
+      </div>
+    </div>
+  </div>
+  <div class="hab-origin--right hab-origin--pkg-list">
+    <p>An <b>origin</b> defines packages that are conceptually related to each other.</p>
+    <p>Automated package builds are enabled once your package has been connected to a GitHub repo that contains a Habitat plan file.</p>
+    <p>You can either <a href="#">upload your packages to the depot</a> and connect them afterwards or connect a repo now and upload your packages later.</p>
+    <p>Read the docs for more information on <a href="#">origins</a>, <a href="#">packages</a>, and the <a href="#">build service</a>.</p>
+  </div>
+</div>
