+++ components/builder-web/app/origin/origin-page/origin-page-routing.module.ts
@@ -1,0 +1,42 @@
+import { NgModule } from "@angular/core";
+import { Routes, RouterModule } from "@angular/router";
+import { OriginPageComponent } from "../origin-page/origin-page.component";
+import { OriginKeysTabComponent } from "./origin-keys-tab/origin-keys-tab.component";
+import { OriginMembersTabComponent } from "./origin-members-tab/origin-members-tab.component";
+import { OriginPackagesTabComponent } from "./origin-packages-tab/origin-packages-tab.component";
+
+const routes: Routes = [
+  {
+    path: "origins/:origin",
+    component: OriginPageComponent,
+    children: [
+      {
+        path: "",
+        redirectTo: "packages",
+        pathMatch: "full"
+      },
+      {
+        path: "packages",
+        component: OriginPackagesTabComponent
+      },
+      {
+        path: "keys",
+        component: OriginKeysTabComponent
+      },
+      {
+        path: "members",
+        component: OriginMembersTabComponent
+      },
+      {
+        path: "**",
+        redirectTo: "packages"
+      }
+    ]
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class OriginPageRoutingModule {}
