+++ components/builder-web/app/origin/origin-page/origin-page.component.scss
@@ -103,72 +103,3 @@
     margin-top: rem(30);
   }
 }
-
-.hab-key-list {
-  &--list {
-    margin-top: $base-spacing;
-  }
-}
-
-.hab-key-add-form {
-  @include shadow-float-more;
-  border-radius: $global-radius;
-  margin-bottom: rem(30);
-  padding: rem(20);
-
-  &--close {
-    float: right;
-  }
-
-  &--label {
-    color: $hab-gray-dark;
-    font-size: rem(16);
-    margin-top: $base-spacing;
-  }
-
-  &--submit {
-    @include outer-container;
-  }
-
-  &--save {
-    @include span-columns(4);
-    margin-top: 0;
-  }
-
-  &--errors {
-    @include span-columns(8);
-    @include omega;
-    color: $hab-red;
-    font-size: $error-font;
-  }
-}
-
-.hab-origin-members-tab {
-  &--section {
-    margin-bottom: rem(60);
-
-    > h3 {
-      margin-bottom: rem(30);
-    }
-
-    &.invite-members > h3 {
-      margin-bottom: rem(10);
-    }
-  }
-
-  &--submit {
-     @include outer-container;
-  }
-
-  &--save {
-    @include span-columns(4);
-    margin-top: 0;
-  }
-
-  &--errors {
-    @include span-columns(8);
-    @include omega;
-    color: $hab-red;
-    font-size: $error-font;
-  }
-}
