+++ components/builder-web/app/origin/origins-page/_origins-page.component.scss
