+++ components/builder-web/app/routes.ts
@@ -17,9 +17,7 @@ import { BuildPageComponent } from "./build-page/build-page.component";
 import { DashboardComponent } from "./dashboard/dashboard.component";
 import { DashboardGuard } from "./dashboard/dashboard.guard";
 import { ExploreComponent } from "./explore/explore.component";
-import { OriginCreatePageComponent } from "./origin-create-page/OriginCreatePageComponent";
-import { OriginPageComponent } from "./origin-page/OriginPageComponent";
-import { OriginsPageComponent } from "./origins-page/OriginsPageComponent";
+import { OriginModule } from "./origin/origin.module";
 import { PackagePageComponent } from "./package-page/PackagePageComponent";
 import { PackagesPageComponent } from "./packages-page/packages-page.component";
 import { ProjectCreatePageComponent } from "./project-create-page/ProjectCreatePageComponent";
@@ -42,18 +40,6 @@ export const routes: Routes = [
     {
         path: "builds/:id",
         component: BuildPageComponent
-    },
-    {
-        path: "origins",
-        component: OriginsPageComponent,
-    },
-    {
-        path: "origins/create",
-        component: OriginCreatePageComponent,
-    },
-    {
-        path: "origins/:origin",
-        component: OriginPageComponent,
     },
     {
         path: "pkgs/search/:query",
