+++ components/builder-web/app/shared/tabs/TabsComponent.ts
